+++ src/test/java/de/adorsys/keycloak/config/service/AuthorizeImportUsingServiceAccountIT.java
@@ -26,23 +26,24 @@ import org.junit.jupiter.api.Order;
 import org.junit.jupiter.api.Test;
 import org.keycloak.representations.idm.ClientRepresentation;
 import org.keycloak.representations.idm.RealmRepresentation;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.context.TestPropertySource;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.Is.is;
 
+//@ClassOrderer(ClassOrderer.OrderAnnotation)
 class AuthorizeImportUsingServiceAccountIT extends AbstractImportTest {
-
-    private static final String REALM_NAME = "service-account";
+    private static final String REALM_NAME = "master";
 
     AuthorizeImportUsingServiceAccountIT() {
         this.resourcePath = "import-files/service-account";
     }
 
     @Test
-    @Order(1)
-    void createClientWithServiceAccountEnabled() {
-        doImport("01_create_realm_client_with_service_account_enabled.json");
+    @Order(0)
+    void createServiceAccountInMasterRealm() {
+        doImport("00_update_realm_create_service_account_in_master_realm.json");
 
         RealmRepresentation realm = keycloakProvider.getInstance().realm(REALM_NAME).toRepresentation();
         assertThat(realm.getRealm(), is(REALM_NAME));
@@ -50,7 +51,7 @@ class AuthorizeImportUsingServiceAccountIT extends AbstractImportTest {
 
         ClientRepresentation client = keycloakRepository.getClient(
                 REALM_NAME,
-                "config-cli"
+                "config-cli-master"
         );
 
         assertThat(client.isServiceAccountsEnabled(), is(true));
@@ -58,24 +59,57 @@ class AuthorizeImportUsingServiceAccountIT extends AbstractImportTest {
 
     @Nested
     @TestPropertySource(properties = {
-            "keycloak.password=invalid",
+            "keycloak.login-realm=service-account",
             "keycloak.grant-type=client_credentials",
             "keycloak.client-id=config-cli",
-            "keycloak.login-realm=service-account",
-            "keycloak.client-secret=config-cli-secret",
             "keycloak.client-secret=config-cli-secret",
     })
-    class importRealmUsingServiceAccount {
+    class importRealmUsingServiceAccountFromDifferentRealm {
+        private static final String REALM_NAME = "service-account";
+
+        @Autowired
+        public RealmImportService realmImportService;
+
         @Test
-        @Order(2)
-        void run() {
-            doImport("02_update_realm_client_with_service_account_enabled.json");
+        void updateExistingRealm() {
+            doImport("02_update_realm_client_with_service_account_enabled.json", realmImportService);
 
             RealmRepresentation realm = keycloakProvider.getInstance().realm(REALM_NAME).toRepresentation();
 
             assertThat(realm.getRealm(), is(REALM_NAME));
             assertThat(realm.isEnabled(), is(true));
             assertThat(realm.getLoginTheme(), is("moped"));
+        }
+    }
+
+    @Nested
+    @TestPropertySource(properties = {
+            "keycloak.login-realm=master",
+            "keycloak.grant-type=client_credentials",
+            "keycloak.client-id=config-cli-master",
+            "keycloak.client-secret=config-cli-master-secret",
+    })
+    class importRealmUsingServiceAccountFromMaster {
+        private static final String REALM_NAME = "service-account";
+
+        @Autowired
+        public RealmImportService realmImportService;
+
+        @Test
+        void createNewRealm() {
+            doImport("01_create_realm_client_with_service_account_enabled.json", realmImportService);
+
+            RealmRepresentation realm = keycloakProvider.getInstance().realm(REALM_NAME).toRepresentation();
+
+            assertThat(realm.getRealm(), is(REALM_NAME));
+            assertThat(realm.isEnabled(), is(true));
+
+            ClientRepresentation client = keycloakRepository.getClient(
+                    REALM_NAME,
+                    "config-cli"
+            );
+
+            assertThat(client.isServiceAccountsEnabled(), is(true));
         }
     }
 }