+++ CHANGELOG.md
@@ -11,6 +11,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Bump keycloak from 12.0.3 to 12.0.4
 
+### Fixed
+
+- Forbidden error while create a new realm with a keycloak service account.
+
 ## [3.1.0] - 2021-02-18
 
 ### Added